feat: use responseEntity

+++ src/main/kotlin/com/microservices/chapter3/CustomerController.kt
@@ -1,8 +1,9 @@
 package com.microservices.chapter3
 
 import org.springframework.beans.factory.annotation.Autowired
+import org.springframework.http.HttpStatus
+import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.*
-import java.util.concurrent.ConcurrentHashMap
 
 @RestController
 class CustomerController {
@@ -10,7 +11,12 @@ class CustomerController {
     private lateinit var customerService: CustomerService
 
     @GetMapping("/customer/{id}")
-    fun getCustomer(@PathVariable id: Int) = customerService.getCustomer(id)
+    fun getCustomer(@PathVariable id: Int): ResponseEntity<Customer?> {
+        val customer = customerService.getCustomer(id)
+        val status = if (customer == null) HttpStatus.NOT_FOUND else HttpStatus.OK
+        return ResponseEntity(customer, status)
+    }
+
 
     @GetMapping("/customers")
     fun getCustomers(@RequestParam(required = false, defaultValue = "") nameFilter: String) {
@@ -18,17 +24,28 @@ class CustomerController {
     }
 
     @PostMapping("/customer")
-    fun createCustomer(@RequestBody customer: Customer) {
+    fun createCustomer(@RequestBody customer: Customer): ResponseEntity<Unit?> {
         customerService.createCustomer(customer)
+        return ResponseEntity<Unit?>(null, HttpStatus.CREATED)
     }
 
     @DeleteMapping("/customer/{id}")
-    fun deleteCustomer(@PathVariable id: Int) {
-        customerService.deleteCustomer(id)
+    fun deleteCustomer(@PathVariable id: Int): ResponseEntity<Unit?> {
+        var status = HttpStatus.NOT_FOUND
+        if (customerService.getCustomer(id) != null) {
+            customerService.deleteCustomer(id)
+            status = HttpStatus.OK
+        }
+        return ResponseEntity<Unit?>(null, status)
     }
 
     @PutMapping("/customer/{id}")
-    fun updateCustomer(@PathVariable id: Int, @RequestBody customer: Customer) {
-        customerService.updateCustomer(id, customer)
+    fun updateCustomer(@PathVariable id: Int, @RequestBody customer: Customer): ResponseEntity<Unit?> {
+        var status = HttpStatus.NOT_FOUND
+        if (customerService.getCustomer(id) != null) {
+            customerService.updateCustomer(id, customer)
+            status = HttpStatus.ACCEPTED
+        }
+        return ResponseEntity<Unit?>(null, status)
     }
 }
